figma_to_angular/requirements.txt
streamlit==1.30.0
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
tiktoken==0.5.2
pillow==10.1.0
numpy==1.24.3
msgpack==1.0.7
anthropic==0.7.8